+++ audiolizer/audiolizer.py
@@ -1,14 +1,14 @@
 # ---
 # jupyter:
 #   jupytext:
-#     formats: py:light
+#     formats: ipynb,py
 #     text_representation:
 #       extension: .py
 #       format_name: light
 #       format_version: '1.5'
 #       jupytext_version: 1.11.3
 #   kernelspec:
-#     display_name: Python 3
+#     display_name: Python 3 (ipykernel)
 #     language: python
 #     name: python3
 # ---
@@ -18,13 +18,11 @@ import numpy as np
 import os
 from plotly.offline import plot
 
-from dash.exceptions import PreventUpdate
-
-from Historic_Crypto import HistoricalData
-
-from Historic_Crypto import Cryptocurrencies
+from history import get_history
 
 # +
+from dash.exceptions import PreventUpdate
+
 from Historic_Crypto import Cryptocurrencies
 
 import flask
@@ -51,126 +49,6 @@ midi_threshold = int(os.environ.get('AUDIOLIZER_MIDI_CACHE_SIZE', 10))
 price_threshold = int(os.environ.get('AUDIOLIZER_PRICE_CACHE_SIZE', 10))
 
 print('cache sizes: \n wav:{}\n midi:{}\n price:{}')
-
-
-def load_date(ticker, granularity, int_):
-    start_ = int_.left.strftime('%Y-%m-%d-%H-%M')
-    end_ = int_.right.strftime('%Y-%m-%d-%H-%M')
-    try:
-        return HistoricalData(ticker,
-                              granularity,
-                              start_,
-                              end_,
-                              ).retrieve_data()
-    except:
-        print('could not load using {} {}'.format(start_, end_))
-        raise
-
-
-def get_gaps(df, granularity):
-    new_ = refactor(df, '{}s'.format(granularity))
-    return new_[new_.close.isna()]
-
-
-# +
-def fetch_data(ticker, granularity, start_, end_):
-    """Need dates in this format %Y-%m-%d-%H-%M"""
-    try:
-        return HistoricalData(ticker,
-                              granularity,
-                              start_,
-                              end_,
-                              ).retrieve_data()
-    except:
-        print('could not load using {} {}'.format(start_, end_))
-        raise
-
-
-def write_data(df, ticker):
-    for t, day in df.groupby(pd.Grouper(freq='1D')):
-        tstr = t.strftime('%Y-%m-%d-%H-%M')
-        fname = audiolizer_temp_dir + '/{}-{}.csv.gz'.format(
-                ticker, t.strftime('%Y-%m-%d'))
-        if len(day) > 1:
-            day.to_csv(fname, compression='gzip')
-            print('wrote {}'.format(fname))
-        
-def fetch_missing(files_status, ticker, granularity):
-    """Iterate over batches of missing dates"""
-    for batch, g in files_status[files_status.found==0].groupby('batch', sort=False):
-        t1, t2 = g.iloc[[0, -1]].index
-        # extend by 1 day whether or not t1 == t2
-        t2 += pd.Timedelta('1D')
-        endpoints = [t.strftime('%Y-%m-%d-%H-%M') for t in [t1, t2]]
-        print('fetching {}, {}'.format(len(g), endpoints))
-        df = fetch_data(ticker, granularity, *endpoints)
-        write_data(df, ticker)
-
-def get_history(ticker, start_date, end_date = None, granularity=granularity):
-    """Fetch/load historical data from Coinbase API at specified granularity
-    
-    Data loaded from start_date through end of end_date
-    params:
-        start_date: (str) (see pandas.to_datetime for acceptable formats)
-        end_date: (str)
-        granularity: (int) seconds (default: 300)
-
-    price data is saved by ticker and date and stored in audiolizer_temp_dir
-    """
-    start_date = pd.to_datetime(start_date).tz_localize(None)
-    
-    today = pd.Timestamp.now().tz_localize(None)
-    if end_date is None:
-        end_date = today + pd.Timedelta('1D')
-    else:
-        end_date = min(today, pd.to_datetime(end_date).tz_localize(None)) + pd.Timedelta('1D')
-        
-    fnames = []
-    foundlings = []
-    dates = []
-    batch = []
-    batch_number = 0
-    last_found = -1
-    for int_ in pd.interval_range(start_date, end_date):
-        dates.append(int_.left)
-        fname = audiolizer_temp_dir + '/{}-{}.csv.gz'.format(
-            ticker, int_.left.strftime('%Y-%m-%d'))
-        found = int(os.path.exists(fname))
-        foundlings.append(found)
-        if found != last_found:
-            batch_number += 1
-        last_found = found
-        batch.append(batch_number)
-        fnames.append(fname)
-    
-    
-    files_status = pd.DataFrame(dict(files=fnames, found=foundlings, batch=batch), index=dates)
-    fetch_missing(files_status, ticker, granularity)
-
-    df = pd.concat(map(lambda file: pd.read_csv(file, index_col='time', parse_dates=True),
-                         fnames)).drop_duplicates()
-    gaps = get_gaps(df, granularity)
-
-    if len(gaps) > 0:
-        print('found {} data gaps'.format(len(gaps)))
-        # fetch the data for each date
-        for start_date in gaps.groupby(pd.Grouper(freq='1d')).first().index:
-            print('\tfetching {}'.format(start_date))
-            int_ = pd.interval_range(start=start_date, periods=1, freq='1d')
-            int_ = pd.Interval(int_.left[0], int_.right[0])
-            int_df = load_date(ticker, granularity, int_)
-            fname = audiolizer_temp_dir + '/{}-{}.csv.gz'.format(
-                ticker, int_.left.strftime('%Y-%m-%d'))
-            int_df.to_csv(fname, compression='gzip')
-
-    df = pd.concat(map(lambda file: pd.read_csv(file,index_col='time', parse_dates=True, compression='gzip'),
-                         fnames)).drop_duplicates()
-
-    return df
-
-
-# new = get_history(ticker, start_date)
-# new
 
 # +
 import audiogen_p3
@@ -581,5 +459,7 @@ if __name__ == '__main__':
         )
 
 # -
+# ! jupytext --sync audiolizer.ipynb
+
